Added a timer class. Improved logger singleton.
Added a timer class that calculates frameTime and keeps track of FPS. It
is to be used for physics calculations involving time.

I also refactored
the logger after learning about a better way to implement singletons. It
makes more sense to me and is safer.

The particles now update their positions based off velocity and
bounce off the walls instead of teleporting on the opposite side.

This will make it easy to implement forces and accelerations.

+++ src/systems/particle_render_system.cpp
@@ -44,7 +44,7 @@ void ParticleRenderSystem::render() {
 	// Bind pipelines and draw here
 	vkCmdBindPipeline(cmd.buffer(), VK_PIPELINE_BIND_POINT_GRAPHICS, _pipeline.pipeline()); // Bind pipeline
 
-	vkCmdBindDescriptorSets(cmd.buffer(), VK_PIPELINE_BIND_POINT_GRAPHICS, _pipeline.pipelineLayout(), 0, _particleSet.size(), _particleSet.data(), 0, nullptr);
+	vkCmdBindDescriptorSets(cmd.buffer(), VK_PIPELINE_BIND_POINT_GRAPHICS, _pipeline.pipelineLayout(), 0, static_cast<uint32_t>(_particleSet.size()), _particleSet.data(), 0, nullptr);
 
 	vkCmdDraw(cmd.buffer(), 6 * _numParticles, 1, 0, 0);
 }
